Add files via upload
Just a recursive python script that produces a lawn bowls draw, if possible, given the number of teams (team_num) and the number of rounds to be played (rnds). The number of available rinks is just team_num/2 and hence the number of teams must be an even number. This is a bare bones tool that outputs match ups on the horizontal axis and rounds on the vertical axis. The tool will try 1 million solutions before giving up. For variety the round robin team draw can be randomised by uncommenting random.schuffle(teams)